requirements.txt
anyio==2.2.0
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colour==0.1.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
entrypoints==0.3
et-xmlfile==1.0.1
geojson==2.5.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.10
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.5
notebook==6.2.0
numpy==1.20.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
pixiedust==1.1.19
pixiedust-node==0.2.5
prometheus-client==0.9.0
prompt-toolkit==3.0.17
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
voila==0.2.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
xlrd==2.0.1
zipp==3.4.0
